refactor (major): refactor DiffGraph functionality into separate services and providers
